+++ demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/androidx/compose/material/icons/outlined/WebStories.kt
@@ -21,7 +21,7 @@ import androidx.compose.material.icons.materialIcon
 import androidx.compose.material.icons.materialPath
 import androidx.compose.ui.graphics.vector.ImageVector
 
-public val Icons.Outlined.WebStories: ImageVector
+internal val Icons.Outlined.WebStories: ImageVector
     get() {
         if (_webStories != null) {
             return _webStories!!